requirements.txt
streamlit
requests
beautifulsoup4
feedparser
pandas
transformers
torch
sentencepiece
python-dotenv
jupyterlab 
openai